Refactor: votingmachine -> consensus.Leader. CollectVotes directly called in Consensus.
